小重构，拆成额外三个类，代码类、代码操作类和静态工具方法类
